third_party/WebKit/LayoutTests/external/csswg-test/css-display-3/display-contents-computed-style-expected.txt

This is a testharness.js-based test.
PASS Serialization of computed style value for display:contents 
PASS display:contents element as inherit parent - explicit inheritance 
PASS display:contents element as inherit parent - implicit inheritance 
FAIL Resolved value should be computed value, not used value, for display:contents assert_equals: expected "auto" but got "588px"
PASS display:contents is blockified for root elements 
Harness: the test ran to completion.
